+++ web/src/App.js
@@ -10,7 +10,6 @@ import Setting from './pages/Setting';
 import EditUser from './pages/User/EditUser';
 import { getLogo, getSystemName } from './helpers';
 import PasswordResetForm from './components/PasswordResetForm';
-import GitHubOAuth from './components/GitHubOAuth';
 import PasswordResetConfirm from './components/PasswordResetConfirm';
 import { UserContext } from './context/User';
 import Channel from './pages/Channel';
@@ -26,7 +25,7 @@ import Midjourney from './pages/Midjourney';
 import Pricing from './pages/Pricing/index.js';
 import Task from "./pages/Task/index.js";
 import Playground from './components/Playground.js';
-import LinuxDoOAuth from './components/LinuxDoOAuth.js';
+import OAuth2Callback from "./components/OAuth2Callback.js";
 
 const Home = lazy(() => import('./pages/Home'));
 const Detail = lazy(() => import('./pages/Detail'));
@@ -178,7 +177,7 @@ function App() {
           path='/oauth/github'
           element={
             <Suspense fallback={<Loading></Loading>}>
-              <GitHubOAuth />
+              <OAuth2Callback type='github'></OAuth2Callback>
             </Suspense>
           }
         />
@@ -186,7 +185,7 @@ function App() {
           path='/oauth/linuxdo'
           element={
             <Suspense fallback={<Loading></Loading>}>
-              <LinuxDoOAuth />
+                <OAuth2Callback type='linuxdo'></OAuth2Callback>
             </Suspense>
           }
         />
